refactor(proc-macro): update imports and dependencies
- Update helper.rs to use new font module structure
- Adjust lib.rs imports after helper crate removal
- Maintain compatibility with existing functionality
